chore(timerProps): renamed timerEventDefinitionProps to timerProps
