+++ src/provider/bpmn/properties/index.js
@@ -8,4 +8,4 @@ export { MessageProps } from './MessageProps';
 export { NameProps } from './NameProps';
 export { ProcessProps } from './ProcessProps';
 export { SignalProps } from './SignalProps';
-export { TimerEventDefinitionProps } from './TimerEventDefinitionProps';
+export { TimerProps } from './TimerProps';
