+++ src/provider/zeebe/properties/index.js
@@ -8,4 +8,4 @@ export { TaskDefinitionProps } from './TaskDefinitionProps';
 export { HeaderProps } from './HeaderProps';
 export { MessageProps } from './MessageProps';
 export { FormProps } from './FormProps';
-export { TimerEventDefinitionProps } from './TimerEventDefinitionProps';
+export { TimerProps } from './TimerProps';
